adds a use case integration test that runs a workflow and validates SLS is in an expected state.
this involves creating two files:

- `spec/integration/someusecase_spec.sh`
- `spec/integration/test_someusecase.tavern.yml`

the files should have the same root name to work properly.

this commit can be used as a template for adding future use cases.

the `_spec.sh` file emulates a human doing some work:

- starts a session (with an existing SLS instance in place)
- imports SLS data
- adds a cabinet using cani
- stops a session and commits the changes to SLS

the `tavern.yml` then checks the state of SLS by the definitions in the
file.  See
https://pykwalify.readthedocs.io/en/master/validation-rules.html for
various ways to check the output.

In this simple example, an ex3000 cabinet was added, so we check for the
x1000 cabinet endpoint and validate it looks the way it should.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>